requirements.txt
# Telegram Bot
aiogram>=3.0.0

# HTTP client (for optional use)
aiohttp>=3.8.0

# Environment variables
python-dotenv>=1.0.0

# Speech-to-Text (FREE, offline)
vosk>=0.3.45

# Audio processing
pydub>=0.25.0

# For handling audio files
ffmpeg-python>=0.2.0

# JSON-RPC for file queue
jsonrpcserver>=5.0.0
